split items state
