+++ tuxemon/event/actions/open_shop.py
@@ -11,7 +11,7 @@ from tuxemon.event.eventaction import EventAction
 from tuxemon.item.economy import Economy
 from tuxemon.npc import NPC
 from tuxemon.states.choice import ChoiceState
-from tuxemon.states.items import ShopBuyMenuState, ShopSellMenuState
+from tuxemon.states.items.shop_menu import ShopBuyMenuState, ShopSellMenuState
 
 
 @final
